more optimizations towards sentry mode

+++ helper_functions/database.py
@@ -13,7 +13,7 @@ class Database:
             connection = sqlite3.connect(file_name)
             connection.execute(f"CREATE TABLE {table_name} (category, item, id INTEGER PRIMARY KEY )")
             connection.commit()
-            return f"A database, {file_name}, has been created."
+            return "A database has been created."
 
     def uploader(self, category, item):
         connection = sqlite3.connect(file_name)
@@ -25,7 +25,7 @@ class Database:
                 return f"Looks like the table: {table_name}, already has the item: {item} in, {category} category"
         connection.execute(f"INSERT INTO {table_name} (category, item) VALUES ('{category}','{item}')")
         connection.commit()
-        return f"I've updated the table: {table_name}, with category: {category}, and item: {item}."
+        return f"I've added the item: {item} to the category: {category}."
 
     def downloader(self):
         connection = sqlite3.connect(file_name)
@@ -40,9 +40,12 @@ class Database:
         connector = connection.cursor()
         connector.execute(f"SELECT category, item from {table_name}")
         response = connector.fetchall()
+        check = 0
         for c, i in response:
-            if i != item and c != item:
-                return f"Looks like there is no item or category with the name: {item}"
+            if i == item or c == item:
+                check += 1
+        if check == 0:
+            return f"Looks like there is no item or category with the name: {item}"
         connection.execute(f"DELETE FROM {table_name} WHERE item='{item}' OR category='{item}'")
         connection.commit()
         return f"Item: {item} has been removed from {table_name}."
